create a hook to request Team Users
