+++ src/pages/TeamOverview.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import {useLocation, useParams} from 'react-router-dom';
+import {useLocation} from 'react-router-dom';
 import {ListItem, UserData} from 'types';
-import {getTeamOverview, getUserData} from '../api';
 import Card from '../components/Card';
 import {Container} from '../components/GlobalComponents';
 import Header from '../components/Header';
 import List from '../components/List';
+import {useTeamUsers} from './hooks/useTeamUsers';
 
 var mapArray = (users: UserData[]) => {
     return users.map(u => {
@@ -54,31 +54,9 @@ var mapTLead = tlead => {
     return <Card columns={columns} url={`/user/${tlead.id}`} navigationProps={tlead} />;
 };
 
-interface PageState {
-    teamLead?: UserData;
-    teamMembers?: UserData[];
-}
-
 const TeamOverview = () => {
     const location = useLocation();
-    const {teamId} = useParams();
-    const [pageData, setPageData] = React.useState<PageState>({});
-    const [isLoading, setIsLoading] = React.useState<boolean>(true);
-
-    React.useEffect(() => {
-        var getTeamUsers = async () => {
-            const {teamLeadId, teamMemberIds = []} = await getTeamOverview(teamId);
-            const [teamLead, ...teamMembers] = await Promise.all([getUserData(teamLeadId), ...teamMemberIds.map(getUserData)]);
-            setPageData(() => {
-                setIsLoading(false);
-                return {
-                    teamLead,
-                    teamMembers,
-                };
-            });
-        };
-        getTeamUsers();
-    }, [teamId]);
+    const {isLoading, pageData} = useTeamUsers();
 
     return (
         <Container>
